Fixed the trunk/deck menu bugs
Of course the text just had to be part of an assembly routine. So I had it jump into the fixed bank, queue up the extended names for the two menus, and then jump back, correcting the whitespace printed afterwards too. The routines are shared between the various trunk/deck menus, thankfully.